Refactor Plan cards to use a common Ability class for their counter threshold flow
