+++ Mage.Sets/src/mage/cards/c/ConstructACosmicCube.java
@@ -1,25 +1,18 @@
 package mage.cards.c;
 
-import java.util.Optional;
 import java.util.UUID;
 import mage.abilities.Ability;
-import mage.abilities.TriggeredAbilityImpl;
 import mage.abilities.common.DrawNthCardTriggeredAbility;
+import mage.abilities.common.PlanCounterThresholdTriggeredAbility;
 import mage.abilities.common.delayed.ReflexiveTriggeredAbility;
-import mage.abilities.costs.common.SacrificeSourceCost;
 import mage.abilities.effects.common.CreateTokenEffect;
-import mage.abilities.effects.common.DoWhenCostPaid;
 import mage.abilities.effects.common.counter.AddCountersSourceEffect;
 import mage.abilities.effects.common.turn.ControlTargetPlayerNextTurnEffect;
 import mage.cards.CardImpl;
 import mage.cards.CardSetInfo;
 import mage.constants.CardType;
 import mage.constants.SubType;
-import mage.constants.Zone;
 import mage.counters.CounterType;
-import mage.game.Game;
-import mage.game.events.GameEvent;
-import mage.game.permanent.Permanent;
 import mage.game.permanent.token.VillainToken;
 import mage.target.common.TargetOpponent;
 
@@ -40,7 +33,12 @@ public final class ConstructACosmicCube extends CardImpl {
         this.addAbility(ability);
 
         // When the seventh plan counter is put on this enchantment, sacrifice it. When you do, you control target opponent during their next turn.
-        this.addAbility(new ConstructACosmicCubeCounterTriggeredAbility());
+        ReflexiveTriggeredAbility reflexive = new ReflexiveTriggeredAbility(
+                new ControlTargetPlayerNextTurnEffect(), false,
+                "you control target opponent during their next turn"
+        );
+        reflexive.addTarget(new TargetOpponent());
+        this.addAbility(new PlanCounterThresholdTriggeredAbility(7, reflexive));
     }
 
     private ConstructACosmicCube(final ConstructACosmicCube card) {
@@ -50,54 +48,5 @@ public final class ConstructACosmicCube extends CardImpl {
     @Override
     public ConstructACosmicCube copy() {
         return new ConstructACosmicCube(this);
-    }
-}
-
-class ConstructACosmicCubeCounterTriggeredAbility extends TriggeredAbilityImpl {
-
-    private static DoWhenCostPaid makeEffect() {
-        ReflexiveTriggeredAbility reflexive = new ReflexiveTriggeredAbility(
-            new ControlTargetPlayerNextTurnEffect(), false,
-            "you control target opponent during their next turn"
-        );
-        reflexive.addTarget(new TargetOpponent());
-        return new DoWhenCostPaid(reflexive, new SacrificeSourceCost().setText("sacrifice it"), "Sacrifice {this}?", false);
-    }
-
-    ConstructACosmicCubeCounterTriggeredAbility() {
-        super(Zone.ALL, makeEffect(), false);
-        setTriggerPhrase("When the seventh plan counter is put on {this}, ");
-    }
-
-    private ConstructACosmicCubeCounterTriggeredAbility(final ConstructACosmicCubeCounterTriggeredAbility ability) {
-        super(ability);
-    }
-
-    @Override
-    public ConstructACosmicCubeCounterTriggeredAbility copy() {
-        return new ConstructACosmicCubeCounterTriggeredAbility(this);
-    }
-
-    @Override
-    public boolean checkEventType(GameEvent event, Game game) {
-        return event.getType() == GameEvent.EventType.COUNTER_ADDED;
-    }
-
-    @Override
-    public boolean checkTrigger(GameEvent event, Game game) {
-        if (!event.getTargetId().equals(getSourceId()) || !event.getData().equals(CounterType.PLAN.getName())) {
-            return false;
-        }
-        int amountAdded = event.getAmount();
-        Permanent sourcePermanent = Optional
-                .ofNullable(game.getPermanent(getSourceId()))
-                .orElse(game.getPermanentEntering(getSourceId()));
-        int planCounters;
-        if (sourcePermanent != null) {
-            planCounters = sourcePermanent.getCounters(game).getCount(CounterType.PLAN);
-        } else {
-            planCounters = amountAdded;
-        }
-        return planCounters - amountAdded < 7 && 7 <= planCounters;
     }
 }
