+++ Mage.Sets/src/mage/cards/d/DoomReignsSupreme.java
@@ -1,14 +1,11 @@
 package mage.cards.d;
 
-import java.util.Optional;
 import java.util.UUID;
 import mage.abilities.Ability;
-import mage.abilities.TriggeredAbilityImpl;
 import mage.abilities.common.EntersBattlefieldAllTriggeredAbility;
+import mage.abilities.common.PlanCounterThresholdTriggeredAbility;
 import mage.abilities.common.delayed.ReflexiveTriggeredAbility;
-import mage.abilities.costs.common.SacrificeSourceCost;
 import mage.abilities.effects.OneShotEffect;
-import mage.abilities.effects.common.DoWhenCostPaid;
 import mage.abilities.effects.common.GainLifeEffect;
 import mage.abilities.effects.common.LoseLifeOpponentsEffect;
 import mage.abilities.effects.common.counter.AddCountersSourceEffect;
@@ -25,8 +22,6 @@ import mage.filter.FilterPermanent;
 import mage.filter.StaticFilters;
 import mage.filter.common.FilterControlledPermanent;
 import mage.game.Game;
-import mage.game.events.GameEvent;
-import mage.game.permanent.Permanent;
 import mage.players.Player;
 import mage.target.common.TargetOpponent;
 import mage.util.CardUtil;
@@ -51,7 +46,13 @@ public final class DoomReignsSupreme extends CardImpl {
 
         // When the fifth plan counter is put on this enchantment, sacrifice it. When you do, target opponent exiles the top five cards of
         // their library. You may cast up to two spells from among the exiled cards without paying their mana costs.
-        this.addAbility(new DoomReignsSupremeCounterTriggeredAbility());
+        ReflexiveTriggeredAbility reflexive = new ReflexiveTriggeredAbility(
+                new DoomReignsSupremeExileEffect(), false,
+                "target opponent exiles the top five cards of their library. "
+                + "You may cast up to two spells from among the exiled cards without paying their mana costs"
+        );
+        reflexive.addTarget(new TargetOpponent());
+        this.addAbility(new PlanCounterThresholdTriggeredAbility(5, reflexive));
     }
 
     private DoomReignsSupreme(final DoomReignsSupreme card) {
@@ -61,56 +62,6 @@ public final class DoomReignsSupreme extends CardImpl {
     @Override
     public DoomReignsSupreme copy() {
         return new DoomReignsSupreme(this);
-    }
-}
-
-class DoomReignsSupremeCounterTriggeredAbility extends TriggeredAbilityImpl {
-
-    private static DoWhenCostPaid makeEffect() {
-        ReflexiveTriggeredAbility reflexive = new ReflexiveTriggeredAbility(
-            new DoomReignsSupremeExileEffect(), false,
-            "target opponent exiles the top five cards of their library. "
-            + "You may cast up to two spells from among the exiled cards without paying their mana costs"
-        );
-        reflexive.addTarget(new TargetOpponent());
-        return new DoWhenCostPaid(reflexive, new SacrificeSourceCost().setText("sacrifice it"), "Sacrifice {this}?", false);
-    }
-
-    DoomReignsSupremeCounterTriggeredAbility() {
-        super(Zone.ALL, makeEffect(), false);
-        setTriggerPhrase("When the fifth plan counter is put on {this}, ");
-    }
-
-    private DoomReignsSupremeCounterTriggeredAbility(final DoomReignsSupremeCounterTriggeredAbility ability) {
-        super(ability);
-    }
-
-    @Override
-    public DoomReignsSupremeCounterTriggeredAbility copy() {
-        return new DoomReignsSupremeCounterTriggeredAbility(this);
-    }
-
-    @Override
-    public boolean checkEventType(GameEvent event, Game game) {
-        return event.getType() == GameEvent.EventType.COUNTER_ADDED;
-    }
-
-    @Override
-    public boolean checkTrigger(GameEvent event, Game game) {
-        if (!event.getTargetId().equals(getSourceId()) || !event.getData().equals(CounterType.PLAN.getName())) {
-            return false;
-        }
-        int amountAdded = event.getAmount();
-        Permanent sourcePermanent = Optional
-            .ofNullable(game.getPermanent(getSourceId()))
-            .orElse(game.getPermanentEntering(getSourceId()));
-        int planCounters;
-        if (sourcePermanent != null) {
-            planCounters = sourcePermanent.getCounters(game).getCount(CounterType.PLAN);
-        } else {
-            planCounters = amountAdded;
-        }
-        return planCounters - amountAdded < 5 && 5 <= planCounters;
     }
 }
